+++ WPF.Net.Examples/Views/WindowMain.xaml.cs
@@ -1,11 +1,15 @@
-﻿using System.Windows;
+﻿using MahApps.Metro.Controls;
+using System;
+using System.Diagnostics;
+using System.Threading.Tasks;
+using System.Windows;
 using WPF.Net.Examples.Models;
 using WPF.Net.Examples.ViewModels;
 // ReSharper disable UnusedMember.Global
 
 namespace WPF.Net.Examples.Views
 {
-    public partial class WindowMain
+    public partial class WindowMain : MetroWindow
     {
         #region Private fields and properties
 
@@ -27,6 +31,8 @@ namespace WPF.Net.Examples.Views
         private void WindowMain_Loaded(object sender, RoutedEventArgs e)
         {
             _appSet = ViewModels.Utils.GetSettings(this);
+            _appSet.WindowMain = this;
+            _appSet.DefaultTheme();
         }
 
         private void ButtonWebClient_OnClick(object sender, RoutedEventArgs e)
@@ -43,7 +49,7 @@ namespace WPF.Net.Examples.Views
         {
             _appSet.ActivePage = Enums.WpfPage.Proxy;
         }
-        
+
 
         private void ButtonPing_OnClick(object sender, RoutedEventArgs e)
         {
@@ -54,12 +60,35 @@ namespace WPF.Net.Examples.Views
         {
             _appSet.ActivePage = Enums.WpfPage.WebRequest;
         }
-        
+
         #endregion
 
         private void ButtonChangelog_OnClick(object sender, RoutedEventArgs e)
         {
             _appSet.ActivePage = Enums.WpfPage.Changelog;
+        }
+
+        private void LaunchGitHubSite(object sender, RoutedEventArgs e)
+        {
+            Task.Run(async () =>
+            {
+                await Task.Delay(TimeSpan.FromMilliseconds(1)).ConfigureAwait(false);
+                Process.Start("https://github.com/DamianMorozov/Net.Examples");
+            });
+        }
+
+        private void LaunchNuGetSite(object sender, RoutedEventArgs e)
+        {
+            Task.Run(async () =>
+            {
+                await Task.Delay(TimeSpan.FromMilliseconds(1)).ConfigureAwait(false);
+                Process.Start("https://www.nuget.org/packages/Net.Utils/");
+            });
+        }
+
+        private void buttonTheme_OnClick(object sender, RoutedEventArgs e)
+        {
+            _appSet.ActivePage = Enums.WpfPage.AppTheme;
         }
     }
 }